refactor: move ValidatedCluster + siblings to controller/mod.rs
