+++ rust/operator-binary/src/controller/validate.rs
@@ -17,11 +17,11 @@ use stackable_operator::{
 use strum::IntoEnumIterator;
 
 use crate::{
-    crd::{HdfsNodeRole, v1alpha1},
-    hdfs_controller::{
-        CONTAINER_IMAGE_BASE_NAME, ValidatedCluster, ValidatedClusterConfig, ValidatedRoleConfig,
-        ValidatedRoleGroupConfig,
+    controller::{
+        ValidatedCluster, ValidatedClusterConfig, ValidatedRoleConfig, ValidatedRoleGroupConfig,
     },
+    crd::{HdfsNodeRole, v1alpha1},
+    hdfs_controller::CONTAINER_IMAGE_BASE_NAME,
     security::opa::HdfsOpaConfig,
 };
